+++ app/models/blacklight_cornell_requests/cul_borrow_direct.rb
@@ -156,18 +156,20 @@ module BlacklightCornellRequests
         json_response = {}
 
         while query_pending
-          response = RestClient.get url
-          if (response.code.to_i == 200)
-            # The ActiveCatalog parameter in the response indicates how many BD catalogs are being
-            # actively searched. When the search is complete, this number should be 0.
-            json_response = JSON.parse(response.body)
-            query_pending = json_response['ActiveCatalog'] > 0
-          elsif (response.code.to_i == 404)
-            # This indicates "no result"
-            query_pending = false
-            return nil
-          else
-            Rails.logger.warn("Warning: Requests unable to complete an item search in Borrow Direct (response: #{response.code} #{response.body}")
+          begin
+            response = RestClient.get url
+            if (response.code.to_i == 200)
+              # The ActiveCatalog parameter in the response indicates how many BD catalogs are being
+              # actively searched. When the search is complete, this number should be 0.
+              json_response = JSON.parse(response.body)
+              query_pending = json_response['ActiveCatalog'] > 0
+            else
+              Rails.logger.warn("Warning: Requests unable to complete an item search in Borrow Direct (response: #{response.code} #{response.body}")
+              query_pending = false
+              return nil
+            end
+          # BD inexplicably returns a 404 if the search query isn't matched.
+          rescue RestClient::NotFound => e
             query_pending = false
             return nil
           end